# term_match_openapi.yaml
openapi: 3.0.3
info:
  title: Term Match API
  description: |
    Term Match API 是基于FAISS + BGE-Large-EN-v1.5模型的英文术语匹配服务，
    支持多粒度匹配（1-3个词）和重叠去重处理。
    
    ## 特性
    - 多文本批量匹配
    - 多粒度N-gram匹配
    - 智能批处理优化
    - 重叠去重处理
    - 本地FAISS索引
    
    ## 模型信息
    - 模型: BGE-Large-EN-v1.5
    - 嵌入维度: 1024
    - 索引类型: FAISS IndexFlatIP
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com

servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.example.com
    description: Production server

paths:
  /term-match/match:
    post:
      summary: 术语匹配
      description: |
        对输入的文本列表进行术语匹配，支持多粒度匹配和重叠去重。
        
        ## 算法流程
        1. 对每个输入文本生成N-grams (1到max_ngram个词)
        2. 批量编码所有N-grams为嵌入向量
        3. 使用FAISS索引搜索相似术语
        4. 按相似度阈值过滤结果
        5. 对重叠匹配进行去重处理
        6. 返回匹配的术语信息
      tags:
        - Term Matching
      parameters:
        - name: similarity_threshold
          in: query
          description: 相似度阈值，低于此值的匹配将被过滤
          required: false
          schema:
            type: number
            format: float
            minimum: 0.0
            maximum: 1.0
            default: 0.7
        - name: top_k
          in: query
          description: 每个文本返回的最大匹配数
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
        - name: max_ngram
          in: query
          description: 最大N-gram长度，支持1-5个词的匹配
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 5
            default: 3
        - name: user_id
          in: query
          description: 限制搜索特定用户的术语
          required: false
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
              description: 待匹配的文本列表
              example: ["machine learning algorithms", "neural network architecture", "deep learning models"]
            examples:
              basic:
                summary: 基础示例
                value: ["machine learning", "neural network"]
              complex:
                summary: 复杂示例
                value: ["machine learning algorithms", "neural network architecture", "deep learning models"]
      responses:
        '200':
          description: 成功返回匹配结果
          content:
            application/json:
              schema:
                type: array
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/TermResponse'
                description: 匹配结果列表，每个输入文本对应一个术语列表
              examples:
                success:
                  summary: 成功匹配示例
                  value:
                    - - term_id: 1
                        en: "machine learning"
                        cn: "机器学习"
                        jp: "機械学習"
                        user_id: 1
                        created_at: "2024-01-01T00:00:00Z"
                        updated_at: "2024-01-01T00:00:00Z"
                      - term_id: 2
                        en: "neural network"
                        cn: "神经网络"
                        jp: "ニューラルネットワーク"
                        user_id: 1
                        created_at: "2024-01-01T00:00:00Z"
                        updated_at: "2024-01-01T00:00:00Z"
                    - - term_id: 2
                        en: "neural network"
                        cn: "神经网络"
                        jp: "ニューラルネットワーク"
                        user_id: 1
                        created_at: "2024-01-01T00:00:00Z"
                        updated_at: "2024-01-01T00:00:00Z"
                    - []
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /term-match/stats:
    get:
      summary: 获取统计信息
      description: 获取术语匹配服务的性能统计信息
      tags:
        - Statistics
      responses:
        '200':
          description: 成功返回统计信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsResponse'
              example:
                total_batches: 15
                total_embeddings: 240
                total_processing_time: 2.345
                total_encoding_time: 1.234
                total_search_time: 0.567
                total_overlap_time: 0.123
                average_batch_size: 16.0
                average_processing_time_per_batch: 0.156
                index_info:
                  total_vectors: 100
                  embedding_dimension: 1024
                  index_type: "IndexFlatIP"

  /term-match/test:
    post:
      summary: 测试接口
      description: 测试术语匹配功能，使用默认参数
      tags:
        - Testing
      parameters:
        - name: text
          in: query
          description: 测试文本
          required: false
          schema:
            type: string
            default: "machine learning algorithms"
        - name: similarity_threshold
          in: query
          description: 相似度阈值
          required: false
          schema:
            type: number
            format: float
            default: 0.6
        - name: top_k
          in: query
          description: 返回结果数量
          required: false
          schema:
            type: integer
            default: 5
      responses:
        '200':
          description: 成功返回测试结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestResponse'
              example:
                input_text: "machine learning algorithms"
                matches:
                  - term_id: 1
                    en: "machine learning"
                    cn: "机器学习"
                    jp: "機械学習"
                    similarity: 0.95
                processing_time: 0.123
                stats:
                  total_ngrams: 3
                  total_embeddings: 3
                  search_time: 0.045

components:
  schemas:
    TermResponse:
      type: object
      properties:
        term_id:
          type: integer
          description: 术语ID
          example: 1
        en:
          type: string
          description: 英文术语
          example: "machine learning"
        cn:
          type: string
          description: 中文翻译
          example: "机器学习"
        jp:
          type: string
          description: 日文翻译
          example: "機械学習"
        user_id:
          type: integer
          description: 用户ID
          example: 1
        created_at:
          type: string
          format: date-time
          description: 创建时间
          example: "2024-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: 更新时间
          example: "2024-01-01T00:00:00Z"
      required:
        - term_id
        - en
        - cn
        - jp

    StatsResponse:
      type: object
      properties:
        total_batches:
          type: integer
          description: 总批处理次数
        total_embeddings:
          type: integer
          description: 总编码向量数
        total_processing_time:
          type: number
          format: float
          description: 总处理时间（秒）
        total_encoding_time:
          type: number
          format: float
          description: 总编码时间（秒）
        total_search_time:
          type: number
          format: float
          description: 总搜索时间（秒）
        total_overlap_time:
          type: number
          format: float
          description: 总重叠处理时间（秒）
        average_batch_size:
          type: number
          format: float
          description: 平均批处理大小
        average_processing_time_per_batch:
          type: number
          format: float
          description: 每批平均处理时间（秒）
        index_info:
          type: object
          properties:
            total_vectors:
              type: integer
              description: 索引中的总向量数
            embedding_dimension:
              type: integer
              description: 嵌入向量维度
            index_type:
              type: string
              description: FAISS索引类型

    TestResponse:
      type: object
      properties:
        input_text:
          type: string
          description: 输入文本
        matches:
          type: array
          items:
            type: object
            properties:
              term_id:
                type: integer
                description: 术语ID
              en:
                type: string
                description: 英文术语
              cn:
                type: string
                description: 中文翻译
              jp:
                type: string
                description: 日文翻译
              similarity:
                type: number
                format: float
                description: 相似度分数
        processing_time:
          type: number
          format: float
          description: 处理时间（秒）
        stats:
          type: object
          properties:
            total_ngrams:
              type: integer
              description: 生成的N-gram数量
            total_embeddings:
              type: integer
              description: 编码的向量数量
            search_time:
              type: number
              format: float
              description: 搜索时间（秒）

    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          description: 错误详情
          example: "Term matching service unavailable: Model path not found"
      required:
        - detail

tags:
  - name: Term Matching
    description: 术语匹配相关接口
  - name: Statistics
    description: 统计信息相关接口
  - name: Testing
    description: 测试相关接口 